+++ src/SearchConsoleClientFactory.php
@@ -33,6 +33,9 @@ class SearchConsoleClientFactory
 
     protected static function configureCache(Google_Client $client, $config)
     {
+        if( is_null($config['store']) ) {
+            return;
+        }
         $config = collect($config);
 
         $store = \Cache::store($config->get('store'));