+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All Notable changes to `laravel-search-console` will be documented in this file.
 
+ - change default cache to null
+
+
 ## [0.1.0] - 2017-09-07
  - add compatibility with Laravel 5.5
  - dropped support for anything lower than Laravel 5.4